+++ fabric/src/main/java/net/mehvahdjukaar/every_compat/modules/fabric/mcaw/MacawFurnitureModule.java
@@ -25,7 +25,7 @@ public class MacawFurnitureModule extends SimpleModule {
     public final SimpleEntrySet<WoodType, Block> bookshelf;
     public final SimpleEntrySet<WoodType, Block> cupboardBookshelf;
     public final SimpleEntrySet<WoodType, Block> strippedBookshelf;
-    public final SimpleEntrySet<WoodType, Block> STRIPPED_CUPBOARD_BOOKSHELF;
+    public final SimpleEntrySet<WoodType, Block> strippedCupboardBookshelf;
 
     //TYPE: CHAIR
     public final SimpleEntrySet<WoodType, Block> chair;
@@ -34,9 +34,9 @@ public class MacawFurnitureModule extends SimpleModule {
     public final SimpleEntrySet<WoodType, Block> stool;
 
     public final SimpleEntrySet<WoodType, Block> strippedChair;
-    public final SimpleEntrySet<WoodType, Block> STRIPPED_MODERN_CHAIR;
-    public final SimpleEntrySet<WoodType, Block> STRIPPED_STRIPED_CHAIR;
-    public final SimpleEntrySet<WoodType, Block> STRIPPED_STOOL;
+    public final SimpleEntrySet<WoodType, Block> strippedModernChair;
+    public final SimpleEntrySet<WoodType, Block> strippedStripedChair;
+    public final SimpleEntrySet<WoodType, Block> strippedStool;
 
     //TYPE: COUNTER
     public final SimpleEntrySet<WoodType, Block> counter;
@@ -45,9 +45,9 @@ public class MacawFurnitureModule extends SimpleModule {
     public final SimpleEntrySet<WoodType, Block> doubleDrawerCounter;
 
     public final SimpleEntrySet<WoodType, Block> strippedCounter;
-    public final SimpleEntrySet<WoodType, Block> STRIPPED_DRAWER_COUNTER;
-    public final SimpleEntrySet<WoodType, Block> STRIPPED_CUPBOARD_COUNTER;
-    public final SimpleEntrySet<WoodType, Block> STRIPPED_DOUBLE_DRAWER_COUNTER;
+    public final SimpleEntrySet<WoodType, Block> strippedDrawerCounter;
+    public final SimpleEntrySet<WoodType, Block> strippedCupboardCounter;
+    public final SimpleEntrySet<WoodType, Block> strippedDoubleDrawerCounter;
 
     //TYPE: DRAWER
     public final SimpleEntrySet<WoodType, Block> drawer;
@@ -56,49 +56,49 @@ public class MacawFurnitureModule extends SimpleModule {
     public final SimpleEntrySet<WoodType, Block> largeDrawer;
     public final SimpleEntrySet<WoodType, Block> lowerBookshelfDrawer;
     public final SimpleEntrySet<WoodType, Block> lowerTripleDrawer;
-    public final SimpleEntrySet<WoodType, Block> TRIPLE_DRAWER;
+    public final SimpleEntrySet<WoodType, Block> tripleDrawer;
 
     public final SimpleEntrySet<WoodType, Block> strippedBookshelfDrawer;
-    public final SimpleEntrySet<WoodType, Block> STRIPPED_DOUBLE_DRAWER;
-    public final SimpleEntrySet<WoodType, Block> STRIPPED_DRAWER;
-    public final SimpleEntrySet<WoodType, Block> STRIPPED_LARGE_DRAWER;
-    public final SimpleEntrySet<WoodType, Block> STRIPPED_LOWER_BOOKSHELF_DRAWER;
-    public final SimpleEntrySet<WoodType, Block> STRIPPED_LOWER_TRIPLE_DRAWER;
-    public final SimpleEntrySet<WoodType, Block> STRIPPED_TRIPLE_DRAWER;
+    public final SimpleEntrySet<WoodType, Block> strippedDoubleDrawer;
+    public final SimpleEntrySet<WoodType, Block> strippedDrawer;
+    public final SimpleEntrySet<WoodType, Block> strippedLargeDrawer;
+    public final SimpleEntrySet<WoodType, Block> strippedLowerBookshelfDrawer;
+    public final SimpleEntrySet<WoodType, Block> strippedLowerTripleDrawer;
+    public final SimpleEntrySet<WoodType, Block> strippedTripleDrawer;
 
     //TYPE: DESK
     public final SimpleEntrySet<WoodType, Block> desk;
     public final SimpleEntrySet<WoodType, Block> coveredDesk;
     public final SimpleEntrySet<WoodType, Block> modernDesk;
 
-    public final SimpleEntrySet<WoodType, Block> STRIPPED_DESK;
-    public final SimpleEntrySet<WoodType, Block> STRIPPED_MODERN_DESK;
-    public final SimpleEntrySet<WoodType, Block> STRIPPED_COVERED_DESK;
+    public final SimpleEntrySet<WoodType, Block> strippedDesk;
+    public final SimpleEntrySet<WoodType, Block> strippedModernDesk;
+    public final SimpleEntrySet<WoodType, Block> strippedCoveredDesk;
 
     //TYPE: TABLE
-    public final SimpleEntrySet<WoodType, Block> TABLE;
+    public final SimpleEntrySet<WoodType, Block> table;
     public final SimpleEntrySet<WoodType, Block> glassTable;
     public final SimpleEntrySet<WoodType, Block> endTable;
     public final SimpleEntrySet<WoodType, Block> coffeeTable;
-    public final SimpleEntrySet<WoodType, Block> STRIPPED_END_TABLE;
-    public final SimpleEntrySet<WoodType, Block> STRIPPED_GLASS_TABLE;
-    public final SimpleEntrySet<WoodType, Block> STRIPPED_TABLE;
+    public final SimpleEntrySet<WoodType, Block> strippedEndTable;
+    public final SimpleEntrySet<WoodType, Block> strippedGlassTable;
+    public final SimpleEntrySet<WoodType, Block> strippedTable;
     public final SimpleEntrySet<WoodType, Block> strippedCoffeeTable;
 
     //TYPE: WARDROBE
-    public final SimpleEntrySet<WoodType, Block> WARDROBE;
+    public final SimpleEntrySet<WoodType, Block> wardrobe;
     public final SimpleEntrySet<WoodType, Block> doubleWardrobe;
     public final SimpleEntrySet<WoodType, Block> modernWardrobe;
-    public final SimpleEntrySet<WoodType, Block> STRIPPED_DOUBLE_WARDROBE;
-    public final SimpleEntrySet<WoodType, Block> STRIPPED_MODERN_WARDROBE;
-    public final SimpleEntrySet<WoodType, Block> STRIPPED_WARDROBE;
+    public final SimpleEntrySet<WoodType, Block> strippedDoubleWardrobe;
+    public final SimpleEntrySet<WoodType, Block> strippedModernWardrobe;
+    public final SimpleEntrySet<WoodType, Block> strippedWardrobe;
 
     public MacawFurnitureModule(String modId) {
         super(modId, "mcfur");
         var tab = MacawsFurniture.FURNITUREGROUP;
 
 
-        WARDROBE = SimpleEntrySet.builder(WoodType.class, "wardrobe",
+        wardrobe = SimpleEntrySet.builder(WoodType.class, "wardrobe",
                         () -> BlockInit.OAK_WARDROBE, () -> WoodTypeRegistry.OAK_TYPE,
                         w -> new TallFurnitureHinge(Utils.copyPropertySafe(w.log)))
                 .requiresChildren("stripped_log")
@@ -109,7 +109,7 @@ public class MacawFurnitureModule extends SimpleModule {
                 .defaultRecipe()
                 .build();
 
-        this.addEntry(WARDROBE);
+        this.addEntry(wardrobe);
 
         modernWardrobe = SimpleEntrySet.builder(WoodType.class, "modern_wardrobe",
                         () -> BlockInit.OAK_MODERN_WARDROBE, () -> WoodTypeRegistry.OAK_TYPE,
@@ -241,7 +241,7 @@ public class MacawFurnitureModule extends SimpleModule {
 
         this.addEntry(lowerTripleDrawer);
 
-        TRIPLE_DRAWER = SimpleEntrySet.builder(WoodType.class, "triple_drawer",
+        tripleDrawer = SimpleEntrySet.builder(WoodType.class, "triple_drawer",
                         () -> BlockInit.OAK_TRIPLE_DRAWER, () -> WoodTypeRegistry.OAK_TYPE,
                         w -> new WideFurniture(Utils.copyPropertySafe(w.log)))
                 .requiresChildren("stripped_log")
@@ -252,7 +252,7 @@ public class MacawFurnitureModule extends SimpleModule {
                 .defaultRecipe()
                 .build();
 
-        this.addEntry(TRIPLE_DRAWER);
+        this.addEntry(tripleDrawer);
 
         desk = SimpleEntrySet.builder(WoodType.class, "desk",
                         () -> BlockInit.OAK_DESK, () -> WoodTypeRegistry.OAK_TYPE,
@@ -287,7 +287,7 @@ public class MacawFurnitureModule extends SimpleModule {
 
         this.addEntry(modernDesk);
 
-        TABLE = SimpleEntrySet.builder(WoodType.class, "table",
+        table = SimpleEntrySet.builder(WoodType.class, "table",
                         () -> BlockInit.OAK_TABLE, () -> WoodTypeRegistry.OAK_TYPE,
                         w -> new TableHitbox(Utils.copyPropertySafe(w.log)))
                 .addTag(BlockTags.MINEABLE_WITH_AXE, Registries.BLOCK)
@@ -296,7 +296,7 @@ public class MacawFurnitureModule extends SimpleModule {
                 .defaultRecipe()
                 .build();
 
-        this.addEntry(TABLE);
+        this.addEntry(table);
 
         endTable = SimpleEntrySet.builder(WoodType.class, "end_table",
                         () -> BlockInit.OAK_END_TABLE, () -> WoodTypeRegistry.OAK_TYPE,
@@ -338,6 +338,7 @@ public class MacawFurnitureModule extends SimpleModule {
                 .addTag(BlockTags.MINEABLE_WITH_AXE, Registries.BLOCK)
                 .addTag(modRes("chair"), Registries.BLOCK)
                 .setTabKey(() -> tab)
+                .requiresChildren("stripped_log")
                 .defaultRecipe()
                 .build();
 
@@ -349,6 +350,7 @@ public class MacawFurnitureModule extends SimpleModule {
                 .addTag(modRes("modern_chair"), Registries.BLOCK)
                 .addTag(BlockTags.MINEABLE_WITH_AXE, Registries.BLOCK)
                 .setTabKey(() -> tab)
+                .requiresChildren("stripped_log")
                 .defaultRecipe()
                 .build();
 
@@ -361,6 +363,7 @@ public class MacawFurnitureModule extends SimpleModule {
                 .addTag(BlockTags.MINEABLE_WITH_AXE, Registries.BLOCK)
                 .setRenderType(() -> RenderType::cutout)
                 .setTabKey(() -> tab)
+                .requiresChildren("stripped_log")
                 .defaultRecipe()
                 .build();
 
@@ -372,6 +375,7 @@ public class MacawFurnitureModule extends SimpleModule {
                 .addTag(BlockTags.MINEABLE_WITH_AXE, Registries.BLOCK)
                 .addTag(modRes("stool_chair"), Registries.BLOCK)
                 .setTabKey(() -> tab)
+                .requiresChildren("stripped_log")
                 .defaultRecipe()
                 .build();
 
@@ -429,7 +433,7 @@ public class MacawFurnitureModule extends SimpleModule {
 
         this.addEntry(cupboardCounter);
 
-        STRIPPED_WARDROBE = SimpleEntrySet.builder(WoodType.class, "wardrobe", "stripped",
+        strippedWardrobe = SimpleEntrySet.builder(WoodType.class, "wardrobe", "stripped",
                         () -> BlockInit.STRIPPED_OAK_WARDROBE, () -> WoodTypeRegistry.OAK_TYPE,
                         w -> new TallFurnitureHinge(Utils.copyPropertySafe(w.log)))
                 .requiresChildren("stripped_log")
@@ -440,9 +444,9 @@ public class MacawFurnitureModule extends SimpleModule {
                 .defaultRecipe()
                 .build();
 
-        this.addEntry(STRIPPED_WARDROBE);
+        this.addEntry(strippedWardrobe);
 
-        STRIPPED_MODERN_WARDROBE = SimpleEntrySet.builder(WoodType.class, "modern_wardrobe", "stripped",
+        strippedModernWardrobe = SimpleEntrySet.builder(WoodType.class, "modern_wardrobe", "stripped",
                         () -> BlockInit.STRIPPED_OAK_MODERN_WARDROBE, () -> WoodTypeRegistry.OAK_TYPE,
                         w -> new TallFurnitureHinge(Utils.copyPropertySafe(w.log)))
                 .requiresChildren("stripped_log")
@@ -453,9 +457,9 @@ public class MacawFurnitureModule extends SimpleModule {
                 .defaultRecipe()
                 .build();
 
-        this.addEntry(STRIPPED_MODERN_WARDROBE);
+        this.addEntry(strippedModernWardrobe);
 
-        STRIPPED_DOUBLE_WARDROBE = SimpleEntrySet.builder(WoodType.class, "double_wardrobe", "stripped",
+        strippedDoubleWardrobe = SimpleEntrySet.builder(WoodType.class, "double_wardrobe", "stripped",
                         () -> BlockInit.STRIPPED_OAK_DOUBLE_WARDROBE, () -> WoodTypeRegistry.OAK_TYPE,
                         w -> new TallFurniture(Utils.copyPropertySafe(w.log)))
                 .requiresChildren("stripped_log")
@@ -466,7 +470,7 @@ public class MacawFurnitureModule extends SimpleModule {
                 .defaultRecipe()
                 .build();
 
-        this.addEntry(STRIPPED_DOUBLE_WARDROBE);
+        this.addEntry(strippedDoubleWardrobe);
 
         strippedBookshelf = SimpleEntrySet.builder(WoodType.class, "bookshelf", "stripped",
                         () -> BlockInit.STRIPPED_OAK_BOOKSHELF, () -> WoodTypeRegistry.OAK_TYPE,
@@ -481,7 +485,7 @@ public class MacawFurnitureModule extends SimpleModule {
 
         this.addEntry(strippedBookshelf);
 
-        STRIPPED_CUPBOARD_BOOKSHELF = SimpleEntrySet.builder(WoodType.class, "bookshelf_cupboard", "stripped",
+        strippedCupboardBookshelf = SimpleEntrySet.builder(WoodType.class, "bookshelf_cupboard", "stripped",
                         () -> BlockInit.STRIPPED_OAK_BOOKSHELF_CUPBOARD, () -> WoodTypeRegistry.OAK_TYPE,
                         w -> new TallFurnitureHinge(Utils.copyPropertySafe(w.log)))
                 .requiresChildren("stripped_log")
@@ -492,9 +496,9 @@ public class MacawFurnitureModule extends SimpleModule {
                 .defaultRecipe()
                 .build();
 
-        this.addEntry(STRIPPED_CUPBOARD_BOOKSHELF);
+        this.addEntry(strippedCupboardBookshelf);
 
-        STRIPPED_DRAWER = SimpleEntrySet.builder(WoodType.class, "drawer", "stripped",
+        strippedDrawer = SimpleEntrySet.builder(WoodType.class, "drawer", "stripped",
                         () -> BlockInit.STRIPPED_OAK_DRAWER, () -> WoodTypeRegistry.OAK_TYPE,
                         w -> new WideFurniture(Utils.copyPropertySafe(w.log)))
                 .requiresChildren("stripped_log")
@@ -505,9 +509,9 @@ public class MacawFurnitureModule extends SimpleModule {
                 .defaultRecipe()
                 .build();
 
-        this.addEntry(STRIPPED_DRAWER);
+        this.addEntry(strippedDrawer);
 
-        STRIPPED_DOUBLE_DRAWER = SimpleEntrySet.builder(WoodType.class, "double_drawer", "stripped",
+        strippedDoubleDrawer = SimpleEntrySet.builder(WoodType.class, "double_drawer", "stripped",
                         () -> BlockInit.STRIPPED_OAK_DOUBLE_DRAWER, () -> WoodTypeRegistry.OAK_TYPE,
                         w -> new WideFurniture(Utils.copyPropertySafe(w.log)))
                 .requiresChildren("stripped_log")
@@ -518,7 +522,7 @@ public class MacawFurnitureModule extends SimpleModule {
                 .defaultRecipe()
                 .build();
 
-        this.addEntry(STRIPPED_DOUBLE_DRAWER);
+        this.addEntry(strippedDoubleDrawer);
 
         strippedBookshelfDrawer = SimpleEntrySet.builder(WoodType.class, "bookshelf_drawer", "stripped",
                         () -> BlockInit.STRIPPED_OAK_BOOKSHELF_DRAWER, () -> WoodTypeRegistry.OAK_TYPE,
@@ -533,7 +537,7 @@ public class MacawFurnitureModule extends SimpleModule {
 
         this.addEntry(strippedBookshelfDrawer);
 
-        STRIPPED_LOWER_BOOKSHELF_DRAWER = SimpleEntrySet.builder(WoodType.class, "lower_bookshelf_drawer", "stripped",
+        strippedLowerBookshelfDrawer = SimpleEntrySet.builder(WoodType.class, "lower_bookshelf_drawer", "stripped",
                         () -> BlockInit.STRIPPED_OAK_LOWER_BOOKSHELF_DRAWER, () -> WoodTypeRegistry.OAK_TYPE,
                         w -> new WideFurniture(Utils.copyPropertySafe(w.log)))
                 .requiresChildren("stripped_log")
@@ -544,9 +548,9 @@ public class MacawFurnitureModule extends SimpleModule {
                 .defaultRecipe()
                 .build();
 
-        this.addEntry(STRIPPED_LOWER_BOOKSHELF_DRAWER);
+        this.addEntry(strippedLowerBookshelfDrawer);
 
-        STRIPPED_LARGE_DRAWER = SimpleEntrySet.builder(WoodType.class, "large_drawer", "stripped",
+        strippedLargeDrawer = SimpleEntrySet.builder(WoodType.class, "large_drawer", "stripped",
                         () -> BlockInit.STRIPPED_OAK_LARGE_DRAWER, () -> WoodTypeRegistry.OAK_TYPE,
                         w -> new WideFurniture(Utils.copyPropertySafe(w.log)))
                 .requiresChildren("stripped_log")
@@ -557,9 +561,9 @@ public class MacawFurnitureModule extends SimpleModule {
                 .defaultRecipe()
                 .build();
 
-        this.addEntry(STRIPPED_LARGE_DRAWER);
+        this.addEntry(strippedLargeDrawer);
 
-        STRIPPED_LOWER_TRIPLE_DRAWER = SimpleEntrySet.builder(WoodType.class, "lower_triple_drawer", "stripped",
+        strippedLowerTripleDrawer = SimpleEntrySet.builder(WoodType.class, "lower_triple_drawer", "stripped",
                         () -> BlockInit.STRIPPED_OAK_LOWER_TRIPLE_DRAWER, () -> WoodTypeRegistry.OAK_TYPE,
                         w -> new WideFurniture(Utils.copyPropertySafe(w.log)))
                 .requiresChildren("stripped_log")
@@ -570,9 +574,9 @@ public class MacawFurnitureModule extends SimpleModule {
                 .defaultRecipe()
                 .build();
 
-        this.addEntry(STRIPPED_LOWER_TRIPLE_DRAWER);
+        this.addEntry(strippedLowerTripleDrawer);
 
-        STRIPPED_TRIPLE_DRAWER = SimpleEntrySet.builder(WoodType.class, "triple_drawer", "stripped",
+        strippedTripleDrawer = SimpleEntrySet.builder(WoodType.class, "triple_drawer", "stripped",
                         () -> BlockInit.STRIPPED_OAK_TRIPLE_DRAWER, () -> WoodTypeRegistry.OAK_TYPE,
                         w -> new WideFurniture(Utils.copyPropertySafe(w.log)))
                 .requiresChildren("stripped_log")
@@ -583,9 +587,9 @@ public class MacawFurnitureModule extends SimpleModule {
                 .defaultRecipe()
                 .build();
 
-        this.addEntry(STRIPPED_TRIPLE_DRAWER);
+        this.addEntry(strippedTripleDrawer);
 
-        STRIPPED_DESK = SimpleEntrySet.builder(WoodType.class, "desk", "stripped",
+        strippedDesk = SimpleEntrySet.builder(WoodType.class, "desk", "stripped",
                         () -> BlockInit.STRIPPED_OAK_DESK, () -> WoodTypeRegistry.OAK_TYPE,
                         w -> new Desk(Utils.copyPropertySafe(w.log)))
                 .requiresChildren("stripped_log")
@@ -595,9 +599,9 @@ public class MacawFurnitureModule extends SimpleModule {
                 .defaultRecipe()
                 .build();
 
-        this.addEntry(STRIPPED_DESK);
+        this.addEntry(strippedDesk);
 
-        STRIPPED_COVERED_DESK = SimpleEntrySet.builder(WoodType.class, "covered_desk", "stripped",
+        strippedCoveredDesk = SimpleEntrySet.builder(WoodType.class, "covered_desk", "stripped",
                         () -> BlockInit.STRIPPED_OAK_COVERED_DESK, () -> WoodTypeRegistry.OAK_TYPE,
                         w -> new Desk(Utils.copyPropertySafe(w.log)))
                 .requiresChildren("stripped_log")
@@ -607,9 +611,9 @@ public class MacawFurnitureModule extends SimpleModule {
                 .defaultRecipe()
                 .build();
 
-        this.addEntry(STRIPPED_COVERED_DESK);
+        this.addEntry(strippedCoveredDesk);
 
-        STRIPPED_MODERN_DESK = SimpleEntrySet.builder(WoodType.class, "modern_desk", "stripped",
+        strippedModernDesk = SimpleEntrySet.builder(WoodType.class, "modern_desk", "stripped",
                         () -> BlockInit.STRIPPED_OAK_MODERN_DESK, () -> WoodTypeRegistry.OAK_TYPE,
                         w -> new Desk(Utils.copyPropertySafe(w.log)))
                 .requiresChildren("stripped_log")
@@ -619,9 +623,9 @@ public class MacawFurnitureModule extends SimpleModule {
                 .defaultRecipe()
                 .build();
 
-        this.addEntry(STRIPPED_MODERN_DESK);
+        this.addEntry(strippedModernDesk);
 
-        STRIPPED_TABLE = SimpleEntrySet.builder(WoodType.class, "table", "stripped",
+        strippedTable = SimpleEntrySet.builder(WoodType.class, "table", "stripped",
                         () -> BlockInit.STRIPPED_OAK_TABLE, () -> WoodTypeRegistry.OAK_TYPE,
                         w -> new TableHitbox(Utils.copyPropertySafe(w.log)))
                 .requiresChildren("stripped_log")
@@ -631,9 +635,9 @@ public class MacawFurnitureModule extends SimpleModule {
                 .defaultRecipe()
                 .build();
 
-        this.addEntry(STRIPPED_TABLE);
+        this.addEntry(strippedTable);
 
-        STRIPPED_END_TABLE = SimpleEntrySet.builder(WoodType.class, "end_table", "stripped",
+        strippedEndTable = SimpleEntrySet.builder(WoodType.class, "end_table", "stripped",
                         () -> BlockInit.STRIPPED_OAK_END_TABLE, () -> WoodTypeRegistry.OAK_TYPE,
                         w -> new TableHitbox(Utils.copyPropertySafe(w.log)))
                 .requiresChildren("stripped_log")
@@ -643,7 +647,7 @@ public class MacawFurnitureModule extends SimpleModule {
                 .defaultRecipe()
                 .build();
 
-        this.addEntry(STRIPPED_END_TABLE);
+        this.addEntry(strippedEndTable);
 
         strippedCoffeeTable = SimpleEntrySet.builder(WoodType.class, "coffee_table", "stripped",
                         () -> BlockInit.STRIPPED_OAK_COFFEE_TABLE, () -> WoodTypeRegistry.OAK_TYPE,
@@ -657,7 +661,7 @@ public class MacawFurnitureModule extends SimpleModule {
 
         this.addEntry(strippedCoffeeTable);
 
-        STRIPPED_GLASS_TABLE = SimpleEntrySet.builder(WoodType.class, "glass_table", "stripped",
+        strippedGlassTable = SimpleEntrySet.builder(WoodType.class, "glass_table", "stripped",
                         () -> BlockInit.STRIPPED_OAK_GLASS_TABLE, () -> WoodTypeRegistry.OAK_TYPE,
                         w -> new TableHitbox(Utils.copyPropertySafe(w.log)))
                 .requiresChildren("stripped_log")
@@ -668,7 +672,7 @@ public class MacawFurnitureModule extends SimpleModule {
                 .defaultRecipe()
                 .build();
 
-        this.addEntry(STRIPPED_GLASS_TABLE);
+        this.addEntry(strippedGlassTable);
 
         strippedChair = SimpleEntrySet.builder(WoodType.class, "chair", "stripped",
                         () -> BlockInit.STRIPPED_OAK_CHAIR, () -> WoodTypeRegistry.OAK_TYPE,
@@ -682,18 +686,19 @@ public class MacawFurnitureModule extends SimpleModule {
 
         this.addEntry(strippedChair);
 
-        STRIPPED_MODERN_CHAIR = SimpleEntrySet.builder(WoodType.class, "modern_chair", "stripped",
+        strippedModernChair = SimpleEntrySet.builder(WoodType.class, "modern_chair", "stripped",
                         () -> BlockInit.STRIPPED_OAK_MODERN_CHAIR, () -> WoodTypeRegistry.OAK_TYPE,
                         w -> new ModernChair(Utils.copyPropertySafe(w.log)))
                 .addTag(modRes("modern_chair"), Registries.BLOCK)
+                .requiresChildren("stripped_log")
                 .addTag(BlockTags.MINEABLE_WITH_AXE, Registries.BLOCK)
                 .setTabKey(() -> tab)
                 .defaultRecipe()
                 .build();
 
-        this.addEntry(STRIPPED_MODERN_CHAIR);
+        this.addEntry(strippedModernChair);
 
-        STRIPPED_STRIPED_CHAIR = SimpleEntrySet.builder(WoodType.class, "striped_chair", "stripped",
+        strippedStripedChair = SimpleEntrySet.builder(WoodType.class, "striped_chair", "stripped",
                         () -> BlockInit.STRIPPED_OAK_STRIPED_CHAIR, () -> WoodTypeRegistry.OAK_TYPE,
                         w -> new StripedChair(Utils.copyPropertySafe(w.log)))
                 .requiresChildren("stripped_log")
@@ -704,9 +709,9 @@ public class MacawFurnitureModule extends SimpleModule {
                 .defaultRecipe()
                 .build();
 
-        this.addEntry(STRIPPED_STRIPED_CHAIR);
+        this.addEntry(strippedStripedChair);
 
-        STRIPPED_STOOL = SimpleEntrySet.builder(WoodType.class, "stool_chair", "stripped",
+        strippedStool = SimpleEntrySet.builder(WoodType.class, "stool_chair", "stripped",
                         () -> BlockInit.STRIPPED_OAK_STOOL_CHAIR, () -> WoodTypeRegistry.OAK_TYPE,
                         w -> new Chair(Utils.copyPropertySafe(w.log)))
                 .requiresChildren("stripped_log")
@@ -716,7 +721,7 @@ public class MacawFurnitureModule extends SimpleModule {
                 .defaultRecipe()
                 .build();
 
-        this.addEntry(STRIPPED_STOOL);
+        this.addEntry(strippedStool);
 
         strippedCounter = SimpleEntrySet.builder(WoodType.class, "counter", "stripped",
                         () -> BlockInit.STRIPPED_OAK_COUNTER, () -> WoodTypeRegistry.OAK_TYPE,
@@ -730,7 +735,7 @@ public class MacawFurnitureModule extends SimpleModule {
 
         this.addEntry(strippedCounter);
 
-        STRIPPED_DRAWER_COUNTER = SimpleEntrySet.builder(WoodType.class, "drawer_counter", "stripped",
+        strippedDrawerCounter = SimpleEntrySet.builder(WoodType.class, "drawer_counter", "stripped",
                         () -> BlockInit.STRIPPED_OAK_DRAWER_COUNTER, () -> WoodTypeRegistry.OAK_TYPE,
                         w -> new StorageCounter(Utils.copyPropertySafe(w.log)))
                 .requiresChildren("stripped_log")
@@ -741,9 +746,9 @@ public class MacawFurnitureModule extends SimpleModule {
                 .defaultRecipe()
                 .build();
 
-        this.addEntry(STRIPPED_DRAWER_COUNTER);
+        this.addEntry(strippedDrawerCounter);
 
-        STRIPPED_DOUBLE_DRAWER_COUNTER = SimpleEntrySet.builder(WoodType.class, "double_drawer_counter", "stripped",
+        strippedDoubleDrawerCounter = SimpleEntrySet.builder(WoodType.class, "double_drawer_counter", "stripped",
                         () -> BlockInit.STRIPPED_OAK_DOUBLE_DRAWER_COUNTER, () -> WoodTypeRegistry.OAK_TYPE,
                         w -> new StorageCounter(Utils.copyPropertySafe(w.log)))
                 .requiresChildren("stripped_log")
@@ -754,9 +759,9 @@ public class MacawFurnitureModule extends SimpleModule {
                 .defaultRecipe()
                 .build();
 
-        this.addEntry(STRIPPED_DOUBLE_DRAWER_COUNTER);
+        this.addEntry(strippedDoubleDrawerCounter);
 
-        STRIPPED_CUPBOARD_COUNTER = SimpleEntrySet.builder(WoodType.class, "cupboard_counter", "stripped",
+        strippedCupboardCounter = SimpleEntrySet.builder(WoodType.class, "cupboard_counter", "stripped",
                         () -> BlockInit.STRIPPED_OAK_CUPBOARD_COUNTER, () -> WoodTypeRegistry.OAK_TYPE,
                         w -> new CupboardCounter(Utils.copyPropertySafe(w.log)))
                 .requiresChildren("stripped_log")
@@ -768,6 +773,6 @@ public class MacawFurnitureModule extends SimpleModule {
                 .defaultRecipe()
                 .build();
 
-        this.addEntry(STRIPPED_CUPBOARD_COUNTER);
+        this.addEntry(strippedCupboardCounter);
     }
 }